BeatWizard/files/requirements.txt
fastapi
uvicorn[standard]
python-multipart
librosa
numpy
scipy
soundfile
supabase==2.3.1
python-dotenv==1.0.1
scikit-learn==1.4.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
tqdm==4.66.1
openai==0.28.1
beautifulsoup4
selenium
requests 